move deprecated column code to proper namespace
