+++ db/tests/queries/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 from sqlalchemy import inspect
-from db.columns.operations.select import get_column_attnum_from_name as get_attnum
+from db.deprecated.columns import get_column_attnum_from_name as get_attnum
 from db.queries.base import DBQuery, InitialColumn, JoinParameter
 from db.metadata import get_empty_metadata
