+++ db/tests/queries/test_base.py
@@ -1,6 +1,6 @@
 from sqlalchemy import inspect
 from db.queries.base import DBQuery, InitialColumn, JoinParameter
-from db.columns.operations.select import get_column_attnum_from_name as get_attnum
+from db.deprecated.columns import get_column_attnum_from_name as get_attnum
 from db.transforms import base as tbase
 from db.metadata import get_empty_metadata
